requirements.txt
ttkbootstrap
Pillow
opencv-python-headless
numpy